Add terms specified by client.
This process was done manually. First we downloaded the spreadsheet and converted
it to a csv file using an online web service. Then we took that csv file and
converted it into JSON using another online web service. Finally, we modified
the controller glossary method to handle the new JSON format.